+++ cishouseholds/pipeline/covid_event_transformations.py
@@ -9,17 +9,23 @@ from cishouseholds.derive import assign_any_symptoms_around_visit
 from cishouseholds.derive import assign_column_value_from_multiple_column_map
 from cishouseholds.derive import assign_date_difference
 from cishouseholds.derive import assign_grouped_variable_from_days_since
+from cishouseholds.derive import assign_last_non_null_value_from_col_list
 from cishouseholds.derive import assign_true_if_any
+from cishouseholds.derive import contact_known_or_suspected_covid_type
 from cishouseholds.derive import count_value_occurrences_in_column_subset_row_wise
 from cishouseholds.edit import conditionally_set_column_values
-# from cishouseholds.edit import fuzzy_update
+from cishouseholds.edit import correct_date_ranges_union_dependent
 from cishouseholds.edit import normalise_think_had_covid_columns
+from cishouseholds.edit import remove_incorrect_dates
 from cishouseholds.edit import update_think_have_covid_symptom_any
 from cishouseholds.edit import update_to_value_if_any_not_null
 from cishouseholds.expressions import all_columns_values_in_list
+from cishouseholds.expressions import any_column_equal_value
 from cishouseholds.expressions import count_occurrence_in_row
 from cishouseholds.impute import fill_forward_event
-from cishouseholds.pipeline.post_union_transformations import derive_contact_any_covid_covid_variables
+from cishouseholds.pipeline.mapping import date_cols_min_date_dict
+
+# from cishouseholds.edit import fuzzy_update
 
 
 def covid_event_transformations(df: DataFrame) -> DataFrame:
@@ -43,10 +49,19 @@ def edit_existing_columns(df: DataFrame) -> DataFrame:
     - contact_known_positive_covid_last_28_days
     - think_had_covid
     - think_had_covid_symptom_*
+    - think_have_covid_onset_date
+    - been_outside_uk_last_return_date
+    - other_covid_infection_test_first_positive_date
+    - other_covid_infection_test_last_negative_date
+    - other_antibody_test_first_positive_date
+    - other_antibody_test_last_negative_date
 
     Reference columns:
     - think_had_covid_admitted_to_hospital
     - think_had_covid_contacted_nhs
+    - visit_datetime
+    - visit_id
+    - participant_id
     """
     invalid_covid_date = "2019-11-17"
 
@@ -116,7 +131,7 @@ def edit_existing_columns(df: DataFrame) -> DataFrame:
         "think_had_covid",
         F.when(conditions.get("think_had_covid_onset_date"), "No").otherwise(F.col("think_had_covid")),
     )
-    think_had_covid_columns = [c for c in df.columns if c.startswith("think_had_covid_symptom_")]
+    # think_had_covid_columns = [c for c in df.columns if c.startswith("think_had_covid_symptom_")]
     # df = fuzzy_update(
     #     df,
     #     id_column="participant_id",
@@ -131,6 +146,25 @@ def edit_existing_columns(df: DataFrame) -> DataFrame:
     #     min_matches=len(think_had_covid_columns),  # num available columns - (4 + date column itself)
     #     filter_out_of_range=True,
     # )
+
+    date_cols_to_correct = [
+        col
+        for col in [
+            "last_covid_contact_date",
+            "last_suspected_covid_contact_date",
+            "think_had_covid_onset_date",
+            "think_have_covid_onset_date",
+            "been_outside_uk_last_return_date",
+            "other_covid_infection_test_first_positive_date",
+            "other_covid_infection_test_last_negative_date",
+            "other_antibody_test_first_positive_date",
+            "other_antibody_test_last_negative_date",
+        ]
+        if col in df.columns
+    ]
+    df = correct_date_ranges_union_dependent(df, date_cols_to_correct, "participant_id", "visit_datetime", "visit_id")
+    df = remove_incorrect_dates(df, date_cols_to_correct, "visit_datetime", "2019-08-01", date_cols_min_date_dict)
+
     return df
 
 
@@ -691,4 +725,44 @@ def data_dependent_derivations(df: DataFrame) -> DataFrame:
         visit_id_column="visit_id",
     )
     df = derive_contact_any_covid_covid_variables(df)
+    return df
+
+
+def derive_contact_any_covid_covid_variables(df: DataFrame) -> DataFrame:
+    """
+    Derive variables related to combination of know and suspected covid data columns.
+    """
+    df = df.withColumn(
+        "contact_known_or_suspected_covid",
+        F.when(
+            any_column_equal_value(
+                ["contact_suspected_positive_covid_last_28_days", "contact_known_positive_covid_last_28_days"], "Yes"
+            ),
+            "Yes",
+        ).otherwise("No"),
+    )
+
+    df = assign_last_non_null_value_from_col_list(
+        df=df,
+        column_name_to_assign="contact_known_or_suspected_covid_latest_date",
+        column_list=["last_covid_contact_date", "last_suspected_covid_contact_date"],
+    )
+
+    df = contact_known_or_suspected_covid_type(
+        df=df,
+        contact_known_covid_type_column="last_covid_contact_type",
+        contact_suspect_covid_type_column="last_suspected_covid_contact_type",
+        contact_any_covid_type_column="contact_known_or_suspected_covid",
+        contact_any_covid_date_column="contact_known_or_suspected_covid_latest_date",
+        contact_known_covid_date_column="last_covid_contact_date",
+        contact_suspect_covid_date_column="last_suspected_covid_contact_date",
+    )
+
+    df = assign_date_difference(
+        df,
+        "contact_known_or_suspected_covid_days_since",
+        "contact_known_or_suspected_covid_latest_date",
+        "visit_datetime",
+    )
+
     return df
