+++ cishouseholds/pipeline/pipeline_stages.py
@@ -3,6 +3,7 @@ from functools import reduce
 from io import BytesIO
 from operator import and_
 from pathlib import Path
+from typing import Any
 from typing import Dict
 from typing import List
 from typing import Optional
@@ -35,7 +36,6 @@ from cishouseholds.hdfs_utils import isdir
 from cishouseholds.hdfs_utils import read_header
 from cishouseholds.hdfs_utils import write_string_to_file
 from cishouseholds.impute import fill_forward_only_to_nulls
-from cishouseholds.impute import post_imputation_wrapper
 from cishouseholds.merge import left_join_keep_right
 from cishouseholds.merge import union_multiple_tables
 from cishouseholds.pipeline.config import get_config
@@ -47,8 +47,6 @@ from cishouseholds.pipeline.generate_outputs import write_csv_rename
 from cishouseholds.pipeline.input_file_processing import extract_input_data
 from cishouseholds.pipeline.input_file_processing import extract_lookup_csv
 from cishouseholds.pipeline.input_file_processing import extract_validate_transform_input_data
-from cishouseholds.pipeline.job_transformations import job_transformations
-from cishouseholds.pipeline.job_transformations import reclassify_work_variables
 from cishouseholds.pipeline.load import add_error_file_log_entry
 from cishouseholds.pipeline.load import check_table_exists
 from cishouseholds.pipeline.load import delete_tables
@@ -63,7 +61,6 @@ from cishouseholds.pipeline.lookup_and_regex_transformations import nims_transfo
 from cishouseholds.pipeline.lookup_and_regex_transformations import ordered_household_id_tranformations
 from cishouseholds.pipeline.lookup_and_regex_transformations import process_healthcare_regex
 from cishouseholds.pipeline.lookup_and_regex_transformations import process_vaccine_regex
-from cishouseholds.pipeline.lookup_and_regex_transformations import replace_design_weights_transformations
 from cishouseholds.pipeline.lookup_and_regex_transformations import transform_cis_soc_data
 from cishouseholds.pipeline.lookup_and_regex_transformations import transform_from_lookups
 from cishouseholds.pipeline.manifest import Manifest
@@ -72,12 +69,10 @@ from cishouseholds.pipeline.mapping import column_name_maps
 from cishouseholds.pipeline.mapping import soc_regex_map
 from cishouseholds.pipeline.post_union_transformations import create_formatted_datetime_string_columns
 from cishouseholds.pipeline.post_union_transformations import get_differences
-from cishouseholds.pipeline.post_union_transformations import impute_key_columns
-from cishouseholds.pipeline.post_union_transformations import post_union_transformations
+from cishouseholds.pipeline.post_union_transformations import union_dependent_derivations
 from cishouseholds.pipeline.reporting import count_variable_option
 from cishouseholds.pipeline.reporting import generate_error_table
 from cishouseholds.pipeline.reporting import generate_lab_report
-from cishouseholds.pipeline.covid_event_transformations import symptom_transformations
 from cishouseholds.pipeline.timestamp_map import csv_datetime_maps
 from cishouseholds.pipeline.validation_calls import validation_ETL
 from cishouseholds.pipeline.validation_schema import soc_schema
@@ -94,6 +89,14 @@ from dummy_data_generation.generate_data import generate_survey_v0_data
 from dummy_data_generation.generate_data import generate_survey_v1_data
 from dummy_data_generation.generate_data import generate_survey_v2_data
 
+# from cishouseholds.impute import post_imputation_wrapper
+
+# from cishouseholds.pipeline.post_union_transformations import fill_forward_events_for_key_columns
+# from cishouseholds.pipeline.post_union_transformations import fill_forwards_transformations
+# from cishouseholds.pipeline.demographic_transformations import impute_key_columns
+# from cishouseholds.pipeline.post_union_transformations import reclassify_work_variables
+# from cishouseholds.pipeline.post_union_transformations import union_dependent_cleaning
+
 # from cishouseholds.pipeline.high_level_transformations import fix_timestamps
 
 pipeline_stages = {}
@@ -108,18 +111,6 @@ def register_pipeline_stage(key):
 
     return _add_pipeline_stage
 
-
-@register_pipeline_stage("job_transformations")
-def execute_job_transformations(input_survey_table:str,output_survey_table:str,soc_lookup_table:str,job_lookup_table:str=None):
-    """"""
-    df = extract_from_table(input_survey_table)
-    soc_lookup_df = extract_from_table(soc_lookup_table)
-    if check_table_exists(job_lookup_table):
-        job_lookup_df = extract_from_table(job_lookup_table,break_lineage=True)
-    df,job_lookup_df = job_transformations(df)
-    update_table(df,output_survey_table,"overwrite")
-    update_table(job_lookup_df,job_lookup_table)
-    return {"output_survey_table":output_survey_table}
 
 @register_pipeline_stage("blind_csv_to_table")
 def blind_csv_to_table(path: str, table_name: str, sep: str = "|"):
@@ -527,7 +518,7 @@ def union_survey_response_files(tables_to_process: List, output_survey_table: st
     df_list = [extract_from_table(table) for table in tables_to_process]
 
     df = union_multiple_tables(df_list)
-    update_table(df, output_survey_table, "overwrite")
+    update_table(df, output_survey_table, "overwrite", survey_table=True)
     return {"output_survey_table": output_survey_table}
 
 
@@ -536,12 +527,14 @@ def join_lookup_table(
     input_survey_table: str,
     output_survey_table: str,
     lookup_table_name: str,
+    join_type: str = "left",
     unjoinable_values: Dict[str, Union[str, int]] = {},
     join_on_columns: List[str] = ["work_main_job_title", "work_main_job_role"],
     lookup_transformations: List[str] = [],
     pre_join_transformations: List[str] = [],
     post_join_transformations: List[str] = [],
     output_table_name_key: str = "output_survey_table",
+    **kwargs: dict,
 ):
     """
     Filters input_survey_table into an unjoinable_df, where all join_on_columns are null, and and applies any
@@ -561,32 +554,37 @@ def join_lookup_table(
     transformations: list
         list of transformation functions to be run on the dataframe once the lookup has been joined
     """
+    transformations_dict: Dict[str, Any]
     transformations_dict = {
         "nims": nims_transformations,
         "blood_past_positive": blood_past_positive_transformations,
         "design_weights_lookup": design_weights_lookup_transformations,
-        "replace_design_weights": replace_design_weights_transformations,
         "ordered_household_id": ordered_household_id_tranformations,
     }
 
     lookup_df = extract_from_table(lookup_table_name)
     for transformation in lookup_transformations:
-        lookup_df = transformations_dict[transformation](lookup_df)
+        lookup_df = transformations_dict[transformation](lookup_df, **kwargs)
 
     df = extract_from_table(input_survey_table)
     for transformation in pre_join_transformations:
-        df = transformations_dict[transformation](df)
+        df = transformations_dict[transformation](df, **kwargs)
 
     unjoinable_df = df.filter(all_columns_null(join_on_columns))
     for col, val in unjoinable_values.items():
         unjoinable_df = unjoinable_df.withColumn(col, F.lit(val))
 
     df = df.filter(any_column_not_null(join_on_columns))
-    df = left_join_keep_right(df, lookup_df, join_on_columns)
+
+    if join_type == "left":
+        df = left_join_keep_right(df, lookup_df, join_on_columns)
+    if join_type == "fullouter":
+        df = df.join(df, lookup_df, how="fullouter", on=join_on_columns)
+
     df = union_multiple_tables([df, unjoinable_df])
 
     for transformation in post_join_transformations:
-        df = transformations_dict[transformation](df)
+        df = transformations_dict[transformation](df, **kwargs)
 
     update_table(df, output_survey_table, "overwrite")
     return {output_table_name_key: output_survey_table}
@@ -692,7 +690,7 @@ def update_vaccine_types(input_survey_table: str, output_survey_table: str, vacc
         )
         df = df.withColumn(vaccine_type_other_col, F.lit(None).cast("string"))
 
-    update_table(df, output_survey_table, "overwrite")
+    update_table(df, output_survey_table, "overwrite", survey_table=True)
     return {"output_survey_table": output_survey_table}
 
 
@@ -773,7 +771,7 @@ def lookup_based_editing(
         )
 
     df = transform_from_lookups(df, cohort_lookup, travel_countries_lookup, tenure_group)
-    update_table(df, output_survey_table, write_mode="overwrite")
+    update_table(df, output_survey_table, write_mode="overwrite", survey_table=True)
     return {"output_survey_table": output_survey_table}
 
 
@@ -789,44 +787,150 @@ def execute_union_dependent_transformations(input_survey_table: str, output_surv
     output_survey_table
     """
     df = extract_from_table(input_survey_table)
-    df = post_union_transformations(df)
-    df = job_transformations(df)
-    df = symptom_transformations(df)
-    update_table(df, output_survey_table, write_mode="overwrite")
+    df = union_dependent_derivations(df).custom_checkpoint()
+    update_table(df, output_survey_table, write_mode="overwrite", survey_table=True)
     return {"output_survey_table": output_survey_table}
 
 
-@register_pipeline_stage("impute_demographic_columns")
-def impute_demographic_columns(input_survey_table: str, imputed_values_table: str, output_survey_table: str):
+# @register_pipeline_stage("fill_forwards_events")
+# def execute_fill_forwards_events(input_survey_table: str, output_survey_table: str):
+#     """
+#     Separates out the fill_forwards_event implementation of last observation carried forwards (LOCF) logic from STATA code
+
+#     Parameters
+#     ----------
+#     input_survey_table
+#     output_survey_table
+#     """
+#     df = extract_from_table(input_survey_table)
+#     df = fill_forward_events_for_key_columns(df)
+#     update_table(df, output_survey_table, write_mode="overwrite", survey_table=True)
+#     return {"output_survey_table": output_survey_table}
+
+
+# @register_pipeline_stage("impute_demographic_columns")
+# def impute_demographic_columns(input_survey_table: str, imputed_values_table: str, output_survey_table: str):
+#     """
+#     Impute values for sex, ethnicity and date of birth.
+#     Assumes that columns to be imputed have been filled forwards, as the latest value from each participant is used.
+#     Specific imputations are carried out for each key demographic column. The resulting columns should have no
+#     missing values.
+#     Stores imputed values in a lookup table, for reuse in subsequent imputation rounds. This table is also backed up
+#     with a datetime suffix.
+#     Also outputs a table of survey response records with imputed values.
+#     Note that this stage depends on geography information from the sample files being available
+#     (from sample file processing).
+
+#     Parameters
+#     ----------
+#     input_survey_table
+#     imputed_values_table
+#         name of HIVE table containing previously imputed values by participant
+#     output_survey_table
+#     """
+#     imputed_value_lookup_df = None
+#     if check_table_exists(imputed_values_table):
+#         imputed_value_lookup_df = extract_from_table(imputed_values_table, break_lineage=True)
+#     df = extract_from_table(input_survey_table)
+
+#     key_columns_imputed_df = impute_key_columns(
+#         df, imputed_value_lookup_df, get_config().get("imputation_log_directory", "./")
+#     )
+#     df_with_imputed_values, new_imputed_value_lookup = post_imputation_wrapper(df, key_columns_imputed_df)
+
+#     update_table(new_imputed_value_lookup, imputed_values_table, "overwrite", archive=True)
+#     update_table(df_with_imputed_values, output_survey_table, "overwrite", survey_table=True)
+#     return {"output_survey_table": output_survey_table}
+
+
+@register_pipeline_stage("geography_and_imputation_dependent_logic")
+def geography_and_imputation_dependent_processing(
+    input_survey_table: str,
+    rural_urban_lookup_path: str,
+    output_survey_table: str,
+):
     """
-    Impute values for sex, ethnicity and date of birth.
-    Assumes that columns to be imputed have been filled forwards, as the latest value from each participant is used.
-    Specific imputations are carried out for each key demographic column. The resulting columns should have no
-    missing values.
-    Stores imputed values in a lookup table, for reuse in subsequent imputation rounds. This table is also backed up
-    with a datetime suffix.
-    Also outputs a table of survey response records with imputed values.
-    Note that this stage depends on geography information from the sample files being available
-    (from sample file processing).
+    Processing that depends on geographies and and imputed demographic infromation.
 
     Parameters
     ----------
     input_survey_table
-    imputed_values_table
-        name of HIVE table containing previously imputed values by participant
-    output_survey_table
+        name of the table containing data to be processed
+    rural_urban_lookup_path
+        path to the rural urban lookup to be joined onto responses
+    edited_table
+        name of table to write processed data to
     """
-    imputed_value_lookup_df = None
-    if check_table_exists(imputed_values_table):
-        imputed_value_lookup_df = extract_from_table(imputed_values_table, break_lineage=True)
     df = extract_from_table(input_survey_table)
-
-    key_columns_imputed_df = impute_key_columns(
-        df, imputed_value_lookup_df, get_config().get("imputation_log_directory", "./")
+    rural_urban_lookup_df = get_or_create_spark_session().read.csv(
+        rural_urban_lookup_path,
+        header=True,
+        schema="""
+            lower_super_output_area_code_11 string,
+            cis_rural_urban_classification string
+        """,
+    )  # Prefer version from sample
+    df = df.join(
+        F.broadcast(rural_urban_lookup_df),
+        how="left",
+        on="lower_super_output_area_code_11",
     )
-    new_imputed_value_lookup = post_imputation_wrapper(df, key_columns_imputed_df)
+    df = assign_outward_postcode(df, "outward_postcode", reference_column="postcode")
 
-    update_table(new_imputed_value_lookup, imputed_values_table, "overwrite", archive=True)
+    df = assign_multigenerational(
+        df=df,
+        column_name_to_assign="multigenerational_household",
+        participant_id_column="participant_id",
+        household_id_column="ons_household_id",
+        visit_date_column="visit_datetime",
+        date_of_birth_column="date_of_birth",
+        country_column="country_name_12",
+    )  # Includes school year and age_at_visit derivations
+
+    df = derive_age_based_columns(df, "age_at_visit")
+    df = assign_age_group_school_year(
+        df,
+        country_column="country_name_12",
+        age_column="age_at_visit",
+        school_year_column="school_year",
+        column_name_to_assign="age_group_school_year",
+    )
+    df = assign_work_patient_facing_now(
+        df,
+        column_name_to_assign="work_patient_facing_now",
+        age_column="age_at_visit",
+        work_healthcare_column="work_health_care_patient_facing",
+    )
+    df = assign_work_person_facing_now(
+        df,
+        column_name_to_assign="work_person_facing_now",
+        work_patient_facing_now_column="work_patient_facing_now",
+        work_social_care_column="work_social_care",
+        age_at_visit_column="age_at_visit",
+    )
+
+    # df = update_work_facing_now_column(
+    #     df,
+    #     "work_patient_facing_now",
+    #     "work_status_v0",
+    #     ["Furloughed (temporarily not working)", "Not working (unemployed, retired, long-term sick etc.)", "Student"],
+    # )
+    # df = reclassify_work_variables(df, spark_session=get_or_create_spark_session(), drop_original_variables=False)
+    df = fill_forward_only_to_nulls(
+        df,
+        id="participant_id",
+        date="visit_datetime",
+        list_fill_forward=[
+            "work_status_v0",
+            "work_status_v1",
+            "work_status_v2",
+            "work_location",
+            "work_not_from_home_days_per_week",
+        ],
+    )
+    df = create_formatted_datetime_string_columns(df)
+    update_table(df, output_survey_table, write_mode="overwrite", survey_table=True)
+    return {"output_survey_table": output_survey_table}
 
 
 @register_pipeline_stage("validate_survey_responses")
@@ -892,7 +996,7 @@ def validate_survey_responses(
     # invalid_survey_responses_table = fix_timestamps(erroneous_survey_responses)
     update_table(validation_check_failures_valid_data_df, valid_validation_failures_table, write_mode="append")
     update_table(validation_check_failures_invalid_data_df, invalid_validation_failures_table, write_mode="append")
-    update_table(valid_survey_responses, output_survey_table, write_mode="overwrite", archive=True)
+    update_table(valid_survey_responses, output_survey_table, write_mode="overwrite", archive=True, survey_table=True)
     update_table(erroneous_survey_responses, invalid_survey_responses_table, write_mode="overwrite")
     return {"output_survey_table": output_survey_table}
 
@@ -1202,7 +1306,6 @@ def calculate_household_level_populations(
     household_level_populations_table
         HIVE table to write household level populations to
     """
-    # TODO: add rural urban lookup to this.
     address_lookup_df = extract_from_table(address_lookup_table).select("unique_property_reference_code", "postcode")
     postcode_lookup_df = (
         extract_from_table(postcode_lookup_table)
