+++ cishouseholds/pipeline/visit_transformations.py
@@ -1,0 +1,58 @@
+import pyspark.sql.functions as F
+from pyspark.sql import DataFrame
+
+from cishouseholds.derive import assign_date_difference
+from cishouseholds.derive import assign_first_visit
+from cishouseholds.derive import assign_last_visit
+from cishouseholds.derive import assign_named_buckets
+from cishouseholds.derive import assign_visit_order
+from cishouseholds.edit import replace_sample_barcode
+
+
+def visit_transformations(df: DataFrame):
+
+    df = assign_visit_order(
+        df=df,
+        column_name_to_assign="visit_order",
+        id="participant_id",
+        order_list=["visit_datetime", "visit_id"],
+    )
+    df = df.withColumn(
+        "participant_visit_status", F.coalesce(F.col("participant_visit_status"), F.col("survey_completion_status"))
+    )
+
+    df = replace_sample_barcode(df=df)
+
+    df = assign_first_visit(
+        df=df,
+        column_name_to_assign="household_first_visit_datetime",
+        id_column="ons_household_id",
+        visit_date_column="visit_datetime",
+    )
+    df = assign_last_visit(
+        df=df,
+        column_name_to_assign="last_attended_visit_datetime",
+        id_column="ons_household_id",
+        visit_status_column="participant_visit_status",
+        visit_date_column="visit_datetime",
+    )
+    df = assign_date_difference(
+        df=df,
+        column_name_to_assign="days_since_enrolment",
+        start_reference_column="household_first_visit_datetime",
+        end_reference_column="visit_datetime",
+    )
+    df = assign_date_difference(
+        df=df,
+        column_name_to_assign="household_weeks_since_survey_enrolment",
+        start_reference_column="survey start",
+        end_reference_column="visit_datetime",
+        format="weeks",
+    )
+    df = assign_named_buckets(
+        df,
+        reference_column="days_since_enrolment",
+        column_name_to_assign="visit_number",
+        map={**{0: 1, 14: 2, 21: 3, 28: 4}, **{i * 28: (i + 3) for i in range(2, 200)}},
+    )
+    pass
