+++ tests/pipeline/test_impute_impute_key_columns.py
@@ -4,7 +4,7 @@ import pytest
 from chispa.dataframe_comparer import assert_df_equality
 from pyspark.sql import functions as F
 
-from cishouseholds.pipeline.post_union_transformations import impute_key_columns
+from cishouseholds.pipeline.demographic_transformations import impute_key_columns
 
 
 @pytest.mark.integration
